Fix issues.

+++ src/app/services/auth.js
@@ -84,6 +84,9 @@ const destroyCurrentToken = async (token) => {
         accesstoken: token,
     })
 
+    if (!dbExists) return
+
+    // eslint-disable-next-line consistent-return
     return dbExists.deleteOne()
 }
